Add systemd unit
The init script uses the su command to invoke irods_control.py as the
irods user. This script performs some checks and then starts
irodsServer.

Using su to drop privileges is not ideal, because su is really designed
for interactive use, to gain privileges; not for automated use to drop
them.

su creates a PAM session for the process that it executes, which can
cause all kinds of weird and wonderful stuff to happen, depending on the
local PAM configuration. A more appropriate tool for use in this
situation would be setpriv or setuidgid. It would be even better if the
process executed by the init script (irods_control.py) took care of
dropping privileges itself. More information on this topic can be found
at <https://jdebp.eu/FGA/dont-abuse-su-for-dropping-privileges.html>.

In particular, when using systemd, su's PAM configuration will likely
invoke pam_systemd.so, which registers a session with logind and
transfers the process that it executes out of the control group that
systemd created for irods.service, and into the control group for the
newly-created session. As a result, systemd loses track of the
processes, and is unable to accurately report whether the service is
still running, or restart it when it exits or is killed.

This commit adds a 'check' mode to irods_control.py, which performs the
same pre-flight checks as 'start', but then exits under the assumption
that a separate service manager (such as systemd) will actaully invoke
irodsServer.

A 'grid_shutdown' mode is also added to irods_control.py, which a
service manager can execute and wait on to shut the service down
cleanly.

Finally, a systemd unit is added, which systemd will use in preference
to the init script. It invokes the 'check' mode mentioned above, and
proceeds to start irodsServer if that exits successfully. As irodsServer
does not make use of forking to indicate readiness, the unit is
configured to launch irodsServer in non-forking mode. In this mode,
systemd treats the irodsServer process that it launched as the 'main'
process of the service; because the default KillMode=control-group is
used, systemd will take care to kill the other processes spawed by the
main process if they are still running if the service exits, or is
killed, or is still running after the 'grid_shutdown' mode mentioned
above is used to request that irodsServer stop gracefully.